Fix the yellow on yellow for warnings
Diagnostic highlight now sets the fg to white and the background to
the appropriate color for the diagnostic severity. Before, we
could end up with a yellow color for warning as background and
a yellow as foreground for a function, which caused unreadable text.

Also fix some more cases where re-render is needed.